requirements.txt
Pillow==5.0.0
requests==2.18.4
